\name{ca1d.erlotinib.totals}
\alias{ca1d.erlotinib.totals}
\title{CA1d cell counts in various concentrations erlotinib}

\description{
CA1d cells expressing a fluorescent nuclear tag (histone H2B-RFP) were plated on a 96-well plate and imaged on a BD Pathway 855 every hour for 72 hours on 11/22/2010. Each column contains the number of cells (expressed in log 2) at each unit of time in the presence of various concentrations of erlotinib.

Information on the CA1d cells is in the following reference: \cite{Fred R. Miller, Herbert D. Soule, Larty Tait, Robert J. Pauley, Sandra R. Wolman, Peter J. Dawson, and Gloria H. Heppner. Xenograft Model of Progressive Human Proliferative Breast Disease.
JNCI J Natl Cancer Inst (1993) 85(21): 1725-1732 doi:10.1093/jnci/85.21.1725}

}

\usage{
  data(ca1d.erlotinib.totals)
}
\value{
   A data frame of CA1d cells counts (in log 2 scale) in the presence of various concentrations of  erlotinib.
}
\author{ Darren Tyson, Shawn Garbett }

\examples{
  data(ca1d.erlotinib.totals)
}




